fix: Force extended length apdu for export object
Basically the SE050 will never respond with more than 256 bytes when responding to a short apdu
This was not an issue previously since we only ever exported keys that were smaller than that.

We cannot really dynamically force the extended apdu based on the key type because this information is
not available in the command.